# packages/backend/oas/openapi.yml
openapi: 3.0.1
info:
  title: revert-api
  version: ''
paths:
...
  /connection/integration-status/{revertPublicToken}:
    get:
      description: >-
        Get the OAuth connection event status. This endpoint responds with a
        `http.ServerResponse` instead of `json`.
      operationId: connection_getIntegrationStatus
      tags:
        - Connection
      parameters:
        - name: revertPublicToken
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetConnectStatusResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/commonBaseError'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/commonBaseError'
...
components:
  schemas:
...
    commonBaseError:
      title: commonBaseError
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GetConnectStatusResponse:
      title: GetConnectStatusResponse
      type: object
      properties:
        publicToken:
          type: string
        status:
          $ref: '#/components/schemas/ConnectionStatus'
        integrationName:
          type: string
        tenantId:
          type: string
        tenantSecretToken:
          type: string
          nullable: true
      required:
        - publicToken
        - status
        - integrationName
        - tenantId
    ConnectionStatus:
      title: ConnectionStatus
      type: string
      enum:
        - SUCCESS
        - FAILED
...
  securitySchemes: {}
servers:
  - url: https://api.revert.dev
    description: Production
  - url: https://api-staging.revert.dev
    description: Staging (This staging environment is helpful for testing!)
  - url: http://localhost:4001
    description: Local